requirements.txt
fastapi
uvicorn[standard]
pydantic
boto3
requests
pytest
python-dotenv
